refactor(type_checker): extract Type impl into type_repr.rs
